cio: Made ReadableProxy load external data lazily
Previously the ReadableProxy immediately fetched external data in the constructor.
This meant that it may have have taken a long time before the actual service was
called. It also meant that we may fetch data we never need. But more importantly,
going forward we may only want to fetch parts of the external data. We now have support
for that in DataProxy, but right now don't use it. We simply delay loading external data into
a local temp file the first time someone calls any of the file related methods.